Delete social-listening.md
Signed-off-by: Matt Germonprez <germonprez@gmail.com>

Removing this file from the Value WG and moving it to the Metrics Model WG. As part of this move, the document has been placed into a Google Doc at: https://docs.google.com/document/d/1ZMuBAQUxKCt765L06FVO4RpjEDbpgIQ6FMfPSuThxCE/edit?usp=sharing	

The reason for putting into a Google Doc is simply to follow the CHAOSS workflow. As the metrics model team takes a look at this model, the process will bring it forward as a released metrics model. 

The reason for the move is that the social listening metric as released by the Value WG is really what we are calling metrics models (models of a collection of atomic metics). From a design perspective, we are keeping such composite metrics organized in the metrics-model WG.